requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
gradio>=4.19.2
sendgrid>=6.11.0
openai>=1.12.0
openai-agents>=0.0.17

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0 